change: リンクリスト, getPostを使わない設定に見直し

+++ app/Plugins/User/Linklists/LinklistsPlugin.php
@@ -32,9 +32,9 @@ class LinklistsPlugin extends UserPluginBase
     /* オブジェクト変数 */
 
     /**
-     * 変更時のPOSTデータ
+     * POST チェックに使用する getPost() 関数を使うか
      */
-    public $post = null;
+    public $use_getpost = false;
 
     /* コアから呼び出す関数 */
 
@@ -51,12 +51,15 @@ class LinklistsPlugin extends UserPluginBase
     }
 
     /**
-     *  権限定義
+     * 権限定義
      */
     public function declareRole()
     {
-        // 権限チェックテーブル (追加チェックなし)
+        // 権限チェックテーブル
         $role_check_table = [];
+        $role_check_table["edit"]        = ['frames.edit'];
+        $role_check_table["save"]        = ['frames.create'];
+        $role_check_table["delete"]      = ['frames.delete'];
         return $role_check_table;
     }
 
@@ -70,23 +73,15 @@ class LinklistsPlugin extends UserPluginBase
         return "editBuckets";
     }
 
-    /**
-     * POST取得関数（コアから呼び出す）
-     * コアがPOSTチェックの際に呼び出す関数
-     */
-    public function getPost($id)
-    {
-        // 一度読んでいれば、そのPOSTを再利用する。
-        if (!empty($this->post)) {
-            return $this->post;
-        }
-
-        // POST を取得する。
-        $this->post = LinklistPost::firstOrNew(['id' => $id]);
-        return $this->post;
-    }
-
     /* private関数 */
+
+    /**
+     * リンクデータ取得
+     */
+    private function getLinklistPost($id)
+    {
+        return LinklistPost::firstOrNew(['id' => $id]);
+    }
 
     /**
      * プラグインのフレーム
@@ -256,7 +251,14 @@ class LinklistsPlugin extends UserPluginBase
     public function edit($request, $page_id, $frame_id, $post_id = null)
     {
         // 記事取得
-        $post = $this->getPost($post_id);
+        $post = $this->getLinklistPost($post_id);
+
+        // 更新時
+        if ($post_id) {
+            if (empty($post->id)) {
+                return $this->view_error("403_inframe", null, 'データ存在チェック');
+            }
+        }
 
         // バケツから linklist_id 取得
         $linklist = $this->getPluginBucket($this->getBucketId());
@@ -276,6 +278,16 @@ class LinklistsPlugin extends UserPluginBase
      */
     public function save($request, $page_id, $frame_id, $post_id = null)
     {
+        // POSTデータのモデル取得
+        $post = $this->getLinklistPost($post_id);
+
+        // 更新時
+        if ($post_id) {
+            if (empty($post->id)) {
+                return $this->view_error("403_inframe", null, 'データ存在チェック');
+            }
+        }
+
         // 項目のエラーチェック
         $validator = Validator::make($request->all(), [
             'title'            => ['required', 'max:255'],
@@ -294,9 +306,6 @@ class LinklistsPlugin extends UserPluginBase
         if ($validator->fails()) {
             return back()->withErrors($validator)->withInput();
         }
-
-        // POSTデータのモデル取得
-        $post = LinklistPost::firstOrNew(['id' => $post_id]);
 
         // バケツから linklist_id 取得
         // bugfix: linklist_idはフレームではなく、表示しているバケツから取得する
@@ -334,6 +343,12 @@ class LinklistsPlugin extends UserPluginBase
     {
         // id がある場合、データを削除
         if ($post_id) {
+
+            $post = $this->getLinklistPost($post_id);
+            if (empty($post->id)) {
+                return $this->view_error("403_inframe", null, 'データ存在チェック');
+            }
+
             // データを削除する。
             LinklistPost::where('id', $post_id)->delete();
         }
